optimize code and add unit test
