+++ src/plugins/workspace/server/utils.test.ts
@@ -5,7 +5,8 @@
 
 import { AuthStatus } from '../../../core/server';
 import { httpServerMock, httpServiceMock } from '../../../core/server/mocks';
-import { generateRandomId, getPrincipalsFromRequest } from './utils';
+import { generateRandomId, getPrincipalsFromRequest, isRequestByDashboardAdmin } from './utils';
+import { getWorkspaceState } from '../../../core/server/utils';
 
 describe('workspace utils', () => {
   const mockAuth = httpServiceMock.createAuth();
@@ -72,5 +73,35 @@ describe('workspace utils', () => {
     expect(() => getPrincipalsFromRequest(mockRequest, mockAuth)).toThrow(
       'UNEXPECTED_AUTHORIZATION_STATUS'
     );
+  });
+
+  it('should be dashboard admin when users match configUsers', () => {
+    const mockRequest = httpServerMock.createOpenSearchDashboardsRequest();
+    const groups: string[] = ['dashboard_admin'];
+    const users: string[] = [];
+    const configGroups: string[] = ['dashboard_admin'];
+    const configUsers: string[] = [];
+    isRequestByDashboardAdmin(mockRequest, groups, users, configGroups, configUsers);
+    expect(getWorkspaceState(mockRequest)?.isDashboardAdmin).toBe(true);
+  });
+
+  it('should be dashboard admin when groups match configGroups', () => {
+    const mockRequest = httpServerMock.createOpenSearchDashboardsRequest();
+    const groups: string[] = [];
+    const users: string[] = ['dashboard_admin'];
+    const configGroups: string[] = [];
+    const configUsers: string[] = ['dashboard_admin'];
+    isRequestByDashboardAdmin(mockRequest, groups, users, configGroups, configUsers);
+    expect(getWorkspaceState(mockRequest)?.isDashboardAdmin).toBe(true);
+  });
+
+  it('should be not dashboard admin when groups do not match configGroups', () => {
+    const mockRequest = httpServerMock.createOpenSearchDashboardsRequest();
+    const groups: string[] = ['dashboard_admin'];
+    const users: string[] = [];
+    const configGroups: string[] = [];
+    const configUsers: string[] = ['dashboard_admin'];
+    isRequestByDashboardAdmin(mockRequest, groups, users, configGroups, configUsers);
+    expect(getWorkspaceState(mockRequest)?.isDashboardAdmin).toBe(false);
   });
 });
